Add exceptions handlers

+++ src/main/java/warehouse/warehousemanagementsystem/product/ProductService.java
@@ -1,8 +1,11 @@
 package warehouse.warehousemanagementsystem.product;
 
 import org.springframework.stereotype.Service;
+import warehouse.warehousemanagementsystem.exception.BadRequestException;
 
+import java.math.BigDecimal;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -16,20 +19,63 @@ public class ProductService {
         return productDao.getAllProducts();
     }
 
-    public Product getProductById(Long id) {
-        return productDao.getProductById(id);
+    public void addProduct(Product product) {
+        if (product.name().isEmpty()
+            || product.price() == null
+            || product.categoryId() == null) {
+            throw new BadRequestException("All fields are required");
+        }
+        if (product.stock() < 0) {
+            throw new BadRequestException("Stock cannot be negative");
+        }
+        if (product.price().compareTo(BigDecimal.valueOf(0)) < 0) {
+            throw new BadRequestException("Price cannot be negative");
+        }
+        if (productDao.getProductByName(product).isPresent()) {
+            throw new BadRequestException("Product already exists");
+        }
+        if (productDao.addProduct(product) != 1) {
+            throw new BadRequestException("Failed to add product");
+        }
     }
 
-    public int addProduct(Product product) {
-        return productDao.addProduct(product);
+    public void deleteProduct(Long id) {
+        Optional<Product> product = productDao.getProductById(id);
+        int result;
+        if (product.isEmpty()) {
+            throw new IllegalArgumentException("Product not found");
+        }
+        try {
+            result = productDao.deleteProduct(id);
+        } catch (Exception e) {
+            throw new IllegalArgumentException("This product is still in use");
+        }
+        if (result != 1) {
+            throw new IllegalArgumentException("Failed to delete product");
+        }
     }
 
-    public int deleteProduct(Long id) {
-        return productDao.deleteProduct(id);
-    }
-
-    public int updateProduct(Product product) {
-        return productDao.updateProduct(product);
+    public void updateProduct(Product product) {
+        if (productDao.getProductById(product.id()).isEmpty()) {
+            throw new BadRequestException("Product not found");
+        }
+        if (productDao.getProductByName(product).isPresent()) {
+            throw new BadRequestException("Product already exists");
+        }
+        if (product.name().isEmpty()
+                || product.price() == null
+                || product.categoryId() == null) {
+            throw new BadRequestException("All fields are required");
+        }
+        if (product.stock() < 0) {
+            throw new BadRequestException("Stock cannot be negative");
+        }
+        if (product.price().compareTo(BigDecimal.valueOf(0)) < 0) {
+            throw new BadRequestException("Price cannot be negative");
+        }
+        if (productDao.updateProduct(product) != 1) {
+            throw new BadRequestException("Failed to update product");
+        }
     }
 
     public List<Product> getProductsByCategory(Long categoryId) {
